Allow ceilometer group permissions to /var/log/...
This updates openstack-ceilometer so that the ceilometer
system group can log to the /var/log/ceilometer directory.

This is required when using the Swift ceilometer middleware
which wants to write its log to
/var/log/ceilometer/swift-proxy-server.log as the swift user.

Standard procedure is that the swift user is somehow going to
be in the ceilometer group. This is how puppet does it
for example:

http://git.openstack.org/cgit/stackforge/puppet-swift/tree/manifests/proxy/ceilometer.pp#n22

Change-Id: I566cbe83cb85604b69c537687b9fafbcb0356543