Seperate entrypoint for tempest plugin
Previously python-ceilometer package installed the entry point for the in-tree
Tempest plugin which cause failure while running tempest list-plugins
because Tempest found the entry point but does not have Tempest plugin
code which is subpackaged in python-ceilometer-tests.
Separate ceilometer_tests egg-info is created with the plugin entry point
so that Tempest includes the plugin only when subpackage is installed.

Change-Id: Ie704e0823732ee3bad30301494a97a0bced5f18a