Move python-oslo-db to common package
In the scenario where Ceilometer is deployed with Gnocchi with
no API, it will be impossible to run the Ceilometer upgrade
binary due to the fact that it is unable to import oslo.db.

Even if oslo.db is not used, it is still imported to parse
the configuration therefore it will generate a traceback
if the API is not installed when the upgrade is attempted.

This patch moves the requirement into the common package and
takes out the requirement from the other subpackages.

Change-Id: Id8e345d8749e013f159f58f138a1354cd4fb7b19