Exclude requests-aws from ceilometer package
requests-aws is needed for RadosGW, and was moved from test-reqs to
runtime requirements [1].
The import of this package is protected [2][3].

Since the import is protected, we can ignore the build requirement.
This avoids the need to include a requirement for a feature that we
don't use.

Added python3-babel, since it is not used in ceilometer since 2020[4],
but is needed to create the translations.

[1] https://opendev.org/openstack/ceilometer/commit/7d199f5751fc3d0e302cbfdf0c11d778edda45eb
[2] https://github.com/openstack/ceilometer/blob/master/ceilometer/objectstore/rgw.py#L97-L104
[3] https://github.com/openstack/ceilometer/commit/c9eb2d44df7cafde1294123d66445ebef4cfb76d
[4] https://review.opendev.org/c/openstack/ceilometer/+/749652

Change-Id: Id675a5a0e99d2016bf824443f7a6a186cf1c8049